+++ page-parse/Mapping/Specific/ElementTitleMapper.cs
@@ -1,0 +1,12 @@
+﻿namespace CSGOStats.Infrastructure.PageParse.Mapping.Specific
+{
+    internal class ElementTitleMapper : ElementAttributeMapper
+    {
+        internal const string ElementTitleValueCode = nameof(ElementTitleMapper);
+
+        public ElementTitleMapper()
+            : base("title")
+        {
+        }
+    }
+}