+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v0.2.1
+## Added
+* Anchor link and image title extractor mappers and theirs registration to default mapper factory.
+* Meaningful errors for markup parse faults.
+## Changed
+* Fixed a bug when property cannot be extracted from markup if container attribute wasn't presented.
+* Fixed a base html value mapper behaviour to be fault tolerant and don't throw key not found exception.
+
 # v0.2.0
 ## Added
 * Content loader abstraction. File/HTTP/raw content load implementation.